backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
python-dateutil==2.9.0
drf-spectacular==0.27.2
psycopg[binary]
python-dotenv
